Extract PipelineDatabaseResources from CalculationContext
