+++ kernel/data-pipeline/core/src/main/java/org/apache/shardingsphere/data/pipeline/core/consistencycheck/table/calculator/CalculationContext.java
@@ -18,113 +18,26 @@
 package org.apache.shardingsphere.data.pipeline.core.consistencycheck.table.calculator;
 
 import lombok.Getter;
-import org.apache.shardingsphere.infra.util.close.QuietlyCloser;
+import org.apache.shardingsphere.data.pipeline.core.ingest.dumper.inventory.query.PipelineDatabaseResources;
 
-import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
 import java.util.Deque;
 import java.util.LinkedList;
-import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.concurrent.atomic.AtomicReference;
 
 /**
  * Calculation context.
  *
  * @param <C> the type of record
  */
+@Getter
 public final class CalculationContext<C> implements AutoCloseable {
     
-    private final AtomicReference<Connection> connection = new AtomicReference<>();
+    private final PipelineDatabaseResources databaseResources = new PipelineDatabaseResources();
     
-    private final AtomicReference<PreparedStatement> preparedStatement = new AtomicReference<>();
-    
-    private final AtomicReference<ResultSet> resultSet = new AtomicReference<>();
-    
-    private final AtomicBoolean databaseResourcesReady = new AtomicBoolean(false);
-    
-    @Getter
     private final Deque<C> recordDeque = new LinkedList<>();
-    
-    /**
-     * Get connection.
-     *
-     * @return connection
-     */
-    public Connection getConnection() {
-        return connection.get();
-    }
-    
-    /**
-     * Set connection.
-     *
-     * @param connection connection
-     */
-    public void setConnection(final Connection connection) {
-        this.connection.set(connection);
-    }
-    
-    /**
-     * Set prepared statement.
-     *
-     * @param preparedStatement prepared statement
-     */
-    public void setPreparedStatement(final PreparedStatement preparedStatement) {
-        this.preparedStatement.set(preparedStatement);
-    }
-    
-    /**
-     * Get result set.
-     *
-     * @return result set
-     */
-    public ResultSet getResultSet() {
-        return resultSet.get();
-    }
-    
-    /**
-     * Set result set.
-     *
-     * @param resultSet result set
-     */
-    public void setResultSet(final ResultSet resultSet) {
-        this.resultSet.set(resultSet);
-    }
-    
-    /**
-     * Check if database resources are ready.
-     *
-     * @return true if database resources are ready, false otherwise
-     */
-    public boolean isDatabaseResourcesReady() {
-        return databaseResourcesReady.get();
-    }
-    
-    /**
-     * Set database resources ready.
-     *
-     * @param databaseResourcesReady true if database resources are ready, false otherwise
-     */
-    public void setDatabaseResourcesReady(final boolean databaseResourcesReady) {
-        this.databaseResourcesReady.set(databaseResourcesReady);
-    }
     
     @Override
     public void close() {
-        resetDatabaseResources();
+        databaseResources.close();
         recordDeque.clear();
-    }
-    
-    /**
-     * Reset database resources.
-     */
-    public void resetDatabaseResources() {
-        setDatabaseResourcesReady(false);
-        QuietlyCloser.close(resultSet.get());
-        QuietlyCloser.close(preparedStatement.get());
-        QuietlyCloser.close(connection.get());
-        resultSet.set(null);
-        preparedStatement.set(null);
-        connection.set(null);
     }
 }
